+++ scripts/macros/classFeatures/cleric/blessedStrikes.js
@@ -1,12 +1,11 @@
+import {constants} from '../../../constants.js';
 import {chris} from '../../../helperFunctions.js';
 import {queue} from '../../../utility/queue.js';
 async function onUse({speaker, actor, token, character, item, args, scope, workflow}) {
-    if (workflow.hitTargets.size === 0 || !workflow.damageRoll) return;
+    if (!workflow.hitTargets.size || !workflow.damageRoll) return;
     let itemType = workflow.item.type;
     if (!(itemType === 'weapon' || (itemType === 'spell' && workflow.castData.castLevel === 0))) return;
-    let effect = chris.findEffect(workflow.actor, 'Blessed Strikes');
-    if (!effect) return;
-    let originItem = await fromUuid(effect.origin);
+    let originItem = chris.getItem(workflow.actor, 'Blessed Strikes');
     if (!originItem) return;
     if (chris.inCombat()) {
         let featureUsed = originItem.flags['chris-premades']?.feature?.blessedStrikes?.used;
@@ -14,7 +13,7 @@ async function onUse({speaker, actor, token, character, item, args, scope, workf
     }
     let queueSetup = await queue.setup(workflow.item.uuid, 'blessedStrikes', 150);
     if (!queueSetup) return;
-    let selection = await chris.dialog('Use Blessed Strikes?', [['Yes', true], ['No', false]]);
+    let selection = await chris.dialog(originItem.name, constants.yesNo, 'Use Blessed Strikes?');
     if (!selection) {
         queue.remove(workflow.item.uuid);
         return;